Greater: Created clientside timeout. Lesser: Changed REJ/raw protocol

+++ src/client/client/send/send.c
@@ -14,7 +14,7 @@ void send_initial_communication(client_t* const client) {
     com_free(&com);
 }
 
-void send_REJ(const client_t* const client, const size_t len, const uint16_t* package_nrs) {
+void send_REJ(const client_t* const client, const size_t len, const uint8_t* package_nrs) {
     com_t com;
     com_init(&com, client->fd, MSG_CONFIRM, client->sock, flags_get_raw(1, FLAG_REJ), 0);
     com.packet->data = malloc(sizeof(uint32_t)+len);
